+++ tests/unit/expressjs/test_app.py
@@ -116,7 +116,6 @@ def test_expressjs_environment_vars(
         state_dir=base_dir / "state",
         service_name=framework_name,
         log_files=[],
-        metrics_target=f"*:{framework_config.port}",
         metrics_path="/metrics",
         unit_name="expressjs/0",
     )